Updated lib/sitemap.js
It's now possible to add an xsl URL the sitemap.xml file via the parameter xslUrl. I edited the file rather quickly, because I needed it fast, but maybe I'm not the only one who likes to link to sitemaps in website footers and likes when even the sitemaps fit the general site design :)

I also weeded out some small typos, replaced this with the var it was already cached in and corrected some false references in method comments.